app/requirements.txt
google-cloud-storage
opencv-python-headless
numpy
functions-framework
cloudevents
